Refactor: SSE 스트리밍 로직 공통 함수로 분리, 타입 파일 분리
