+++ src/features/chatbot/hooks/useChatInitialization.ts
@@ -1,0 +1,3 @@
+const useChatInitialization = () => {};
+
+export default useChatInitialization;
