+++ src/features/comment/types/message.ts
@@ -1,0 +1,5 @@
+export interface IMessage {
+  id: number;
+  type: string;
+  content: string;
+}
